Let context.dpToPx return an Int as Float for pixels do not make sense.

+++ app/src/main/java/in/arunkumarsampath/transitionx/sample/extensions/context.kt
@@ -4,7 +4,7 @@ package `in`.arunkumarsampath.transitionx.sample.extensions
 
 import android.content.Context
 
-inline fun Context.dpToPx(dp: Double): Float {
+inline fun Context.dpToPx(dp: Double): Int {
     val displayMetrics = resources.displayMetrics
-    return (dp * displayMetrics.density + 0.5).toFloat()
+    return (dp * displayMetrics.density + 0.5).toInt()
 }
